episode 20 (retake)

+++ app/Livewire/ArticleList.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -14,8 +15,25 @@ class ArticleList extends AdminComponent
 
     public $showOnlyPublished = false;
 
+    #[Computed/*(persist: true)*/]
+    public function articles() {
+        $query = Article::query();
+
+        if ($this->showOnlyPublished) {
+            $query->where('published', 1);
+        }
+
+        return $query->paginate(10, pageName: 'articles-page');
+    }
+
     public function delete(Article $article) {
+        if ($this->articles->count() < 10) {
+            throw new \Exception("Nope");
+        }
+
         $article->delete();
+        unset($this->articles);
+        cache()->forget('published-count');
     }
 
     public function showAll() {
@@ -26,18 +44,5 @@ class ArticleList extends AdminComponent
     public function showPublished() {
         $this->showOnlyPublished = true;
         $this->resetPage(pageName: 'articles-page');
-    }
-
-    public function render()
-    {
-        $query = Article::query();
-
-        if ($this->showOnlyPublished) {
-            $query->where('published', 1);
-        }
-
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10, pageName: 'articles-page'),
-        ]);
     }
 }
